studyai_web_deployment/requirements.txt
flask
flask-login
flask-sqlalchemy
flask-wtf
gunicorn
python-dotenv
sentence-transformers
langchain
pypdf
pydantic
